updating readme to skip missing files

+++ Makefile
@@ -85,9 +85,9 @@ deb_dist:
 	python3 setup.py --command-packages=stdeb.command sdist_dsc
 
 deb_custom:
-	cp debian/$(REPO_NAME).conf $(wildcard deb_dist/*/debian)/$(REPO_NAME).default
-	cp debian/$(REPO_NAME).postinst $(wildcard deb_dist/*/debian)/$(REPO_NAME).postinst
-	cp debian/$(REPO_NAME).service $(wildcard deb_dist/*/debian)/$(REPO_NAME).service
+	test -f debian/$(REPO_NAME).conf && cp debian/$(REPO_NAME).conf $(wildcard deb_dist/*/debian)/$(REPO_NAME).default || true
+	test -f debian/$(REPO_NAME).postinst && cp debian/$(REPO_NAME).postinst $(wildcard deb_dist/*/debian)/$(REPO_NAME).postinst || true
+	test -f debian/$(REPO_NAME).service && cp debian/$(REPO_NAME).service $(wildcard deb_dist/*/debian)/$(REPO_NAME).service || true
 
 bdist_deb: deb_dist deb_custom
 	cd deb_dist/$(REPO_NAME)-*/ && dpkg-buildpackage -rfakeroot -uc -us
